+++ qom/ui/plotters/PlotterPlotly.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+ 
+"""Class to handle plotly plots."""
+
+__name__    = 'qom.ui.plotters.PlotterPlotly'
+__authors__ = ['Sampreet Kalita']
+__created__ = '2020-10-06'
+__updated__ = '2020-10-06'