+++ qom/ui/plotters/__init__.py
@@ -1,0 +1,1 @@
+from qom.ui.plotters.PlotterMPL import PlotterMPL